refactor(gir): Restructure imports and introduce new record and interface types
- Reorganized import statements to improve clarity and maintainability, moving to specific files for introspected classes and types.
- Introduced new `IntrospectedRecord` and `IntrospectedInterface` classes to enhance type safety and organization.
- Removed the deprecated `class.ts` file, consolidating related functionality into more specific modules.
- Updated various files to utilize the new types, ensuring consistency across the codebase.
