+++ packages/lib/src/gir-module.ts
@@ -30,7 +30,10 @@ import type { GirNSMember } from './gir/namespace.ts'
 import { IntrospectedAlias } from './gir/alias.ts'
 import { IntrospectedBase } from './gir/introspected-base.ts'
 import { IntrospectedNamespaceMember } from './gir/introspected-namespace-member.ts'
-import { IntrospectedBaseClass, IntrospectedClass, IntrospectedRecord, IntrospectedInterface } from './gir/class.ts'
+import { IntrospectedBaseClass } from './gir/introspected-base-class.ts'
+import { IntrospectedClass } from './gir/introspected-class.ts'
+import { IntrospectedRecord } from './gir/record.ts'
+import { IntrospectedInterface } from './gir/interface.ts'
 import { IntrospectedConstant } from './gir/const.ts'
 import { IntrospectedEnum, IntrospectedError } from './gir/enum.ts'
 import {
