+++ packages/lib/src/gir/data.ts
@@ -1,0 +1,12 @@
+export const enum ClassInjectionMember {
+  MEMBER = "member",
+  CONSTRUCTOR = "_constructor",
+  PROPERTY = "prop",
+  FIELD = "field",
+  MAIN_CONSTRUCTOR = "constructor"
+}
+
+export enum FilterBehavior {
+  DELETE,
+  PRESERVE
+}
