+++ packages/lib/src/gir/gir-complex-record.ts
@@ -1,0 +1,7 @@
+import { IntrospectedRecord } from "./record.ts";
+
+export class GirComplexRecord extends IntrospectedRecord {
+  isSimple() {
+      return false;
+  }
+}
