+++ packages/lib/src/gir/index.ts
@@ -1,15 +1,21 @@
 export * from "./alias.ts";
-export * from "./class.ts";
-export * from "./introspected-base.ts";
-export * from "./introspected-namespace-member.ts";
-export * from "./introspected-class-member.ts";
+export * from "./callback.ts";
 export * from "./const.ts";
+export * from "./data.ts";
 export * from "./enum.ts";
 export * from "./function.ts";
 export * from "./callback.ts";
 export * from "./generics.ts";
+export * from "./gir-complex-record.ts";
+export * from "./interface.ts";
+export * from "./introspected-base-class.ts";
+export * from "./introspected-base.ts";
+export * from "./introspected-class-member.ts";
+export * from "./introspected-class.ts";
+export * from "./introspected-namespace-member.ts";
 export * from "./namespace.ts";
 export * from "./promisify.ts";
 export * from "./property.ts";
+export * from "./record.ts";
 export * from "./registry.ts";
 export * from "./signal.ts";
