+++ packages/lib/src/types/any-introspected-type.ts
@@ -1,0 +1,3 @@
+import type { IntrospectedBase } from "../gir/introspected-base.ts";
+
+export type AnyIntrospectedType = IntrospectedBase<any>;
