+++ packages/lib/src/types/index.ts
@@ -1,5 +1,8 @@
 export * from './answer-version.ts'
+export * from './any-introspected-type.ts'
+export * from './class-member.ts'
 export * from './class-parent.ts'
+export * from './class.ts'
 export * from './config-flags.ts'
 export * from './construct-name.ts'
 export * from './dependency-map.ts'
